refactor aux.rs
