+++ chips/cc26x0/src/aux_wuc.rs
